+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 6.12.3 (4/9/2019 - )
 * http: update okHTTP to 3.14.1
 * search: update es to 7.0.0, change index.flush to index.refresh (those 2 are different in ES)
+* sys: sys.properties supports allowCIDRs in 2 formats, 1) cidr,cidr 2) name1: cidr, cidr; name2: cidr
 
 ### 6.12.2 (4/2/2019 - 4/9/2019)
 * http: update undertow to 2.0.20